+++ src/main/java/com/buuz135/simpleclaims/commands/subcommand/chunk/UnclaimChunkCommand.java
@@ -21,6 +21,7 @@ public class UnclaimChunkCommand extends AbstractAsyncCommand {
 
     public UnclaimChunkCommand() {
         super("unclaim", "Unclaims the chunk where you are");
+        this.requirePermission(CommandMessages.BASE_PERM + "unclaim");
     }
 
     @NonNullDecl
@@ -50,6 +51,7 @@ public class UnclaimChunkCommand extends AbstractAsyncCommand {
                             return;
                         }
                         ClaimManager.getInstance().unclaimRawCoords(player.getWorld().getName(), (int) playerRef.getTransform().getPosition().getX(), (int) playerRef.getTransform().getPosition().getZ());
+                        ClaimManager.getInstance().queueMapUpdate(player.getWorld(), chunk.getChunkX(), chunk.getChunkZ());
                         player.sendMessage(CommandMessages.UNCLAIMED);
                     }
                 }, world);