+++ src/main/java/com/buuz135/simpleclaims/commands/subcommand/party/CreatePartyCommand.java
@@ -23,7 +23,7 @@ public class CreatePartyCommand extends AbstractAsyncCommand {
 
     public CreatePartyCommand() {
         super("create", "Creates a new party");
-        this.setPermissionGroup(GameMode.Adventure);
+        this.requirePermission(CommandMessages.BASE_PERM + "create-party");
     }
 
     @NonNullDecl
@@ -43,7 +43,7 @@ public class CreatePartyCommand extends AbstractAsyncCommand {
                         commandContext.sendMessage(CommandMessages.IN_A_PARTY);
                         return;
                     }
-                    party = ClaimManager.getInstance().createParty(player, playerRef);
+                    party = ClaimManager.getInstance().createParty(player, playerRef, false);
                     player.sendMessage(CommandMessages.PARTY_CREATED);
                     player.getPageManager().openCustomPage(ref, store, new PartyInfoEditGui(playerRef, party, false));
                 }, world);